Struct photon_core::App

pub struct App { /* fields omitted */ }

Application context

Methods

impl App

fn delay_ms(&mut self, ms: u32)

Waits for ms milliseconds

During this time, the scheduler may execute cloud functions so no outstanding borrows to global resources may exist when this method is called